@@ -1,6 +1,6 @@
 <?php
 
-namespace atoum\AtoumBundle\DependencyInjection;
+namespace Atoum\AtoumBundle\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
@@ -9,12 +9,9 @@ use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\Config\Definition\Processor;
 
 /**
- * AtoumAtoumExtension
- *
- * @uses Extension
  * @author Stephane PY <py.stephane1@gmail.com>
  */
-class AtoumAtoumExtension extends Extension
+class AtoumExtension extends Extension
 {
     /**
      * {@inheritdoc}
@@ -30,13 +27,5 @@ class AtoumAtoumExtension extends Extension
         $loader->load('configuration.xml');
 
         $container->setParameter('atoum.bundles', $config['bundles']);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAlias()
-    {
-        return 'atoum';
     }
 }
