@@ -1,5 +1,5 @@
 <?php
-namespace atoum\AtoumBundle\DomCrawler;
+namespace Atoum\AtoumBundle\DomCrawler;
 
 use Symfony\Component\DomCrawler;
 
