@@ -1,5 +1,5 @@
 <?php
-namespace atoum\AtoumBundle\tests\units\Test\Units;
+namespace Atoum\AtoumBundle\tests\units\Test\Units;
 
 use mageekguy\atoum;
 
@@ -13,13 +13,13 @@ class Test extends atoum\test
     public function testSetAssertionManager()
     {
         $this
-            ->if($object = new \mock\atoum\AtoumBundle\Test\Units\Test())
+            ->if($object = new \mock\Atoum\AtoumBundle\Test\Units\Test())
             ->and($manager = new \mock\mageekguy\atoum\test\assertion\manager())
             ->then
                 ->object($object->setAssertionManager($manager))->isIdenticalTo($object)
                 ->mock($manager)
                     ->call('setHandler')->withArguments('faker')->once()
-            ->if($object = new \mock\atoum\AtoumBundle\Test\Units\Test())
+            ->if($object = new \mock\Atoum\AtoumBundle\Test\Units\Test())
             ->and($this->calling($object)->getFaker = $generator = new \mock\Faker\Generator())
             ->and($this->calling($generator)->__call->doesNothing())
             ->then
@@ -32,7 +32,7 @@ class Test extends atoum\test
     public function testGetFaker()
     {
         $this
-            ->if($object = new \mock\atoum\AtoumBundle\Test\Units\Test())
+            ->if($object = new \mock\Atoum\AtoumBundle\Test\Units\Test())
             ->then
                 ->object($faker = $object->getFaker())->isInstanceOf('\\Faker\\Generator')
                 ->object($object->getFaker())
